requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-groq
ddgs
